修改混用有符号数和无符号数

+++ frameworks/core/components_ng/pattern/waterflow/layout/sliding_window/water_flow_layout_info_sw.cpp
@@ -577,13 +577,13 @@ bool WaterFlowLayoutInfoSW::AdjustLanes(const std::vector<WaterFlowSections::Sec
         return false;
     }
     const size_t n = sections.size();
-    const size_t curSegIdx = GetSegment(newStartIndex_);
+    const size_t curSegIdx = static_cast<size_t>(GetSegment(newStartIndex_));
     auto curSection = sections[curSegIdx];
     if (curSection.OnlyCountDiff(prevSection)) {
         // move old lanes_[prevSegIdx,...] to Lanes_[curSegIdx,...]
         if (n <= lanes_.size()) {
             // means curSegIdx <= prevSegIdx
-            for (size_t i = start; i < curSegIdx; ++i) {
+            for (size_t i = static_cast<size_t>(start); i < curSegIdx; ++i) {
                 lanes_[i] = std::vector<Lane>(sections[i].crossCount.value_or(1));
             }
             for (size_t i = curSegIdx; i < n; ++i) {
@@ -597,7 +597,7 @@ bool WaterFlowLayoutInfoSW::AdjustLanes(const std::vector<WaterFlowSections::Sec
             for (size_t i = n - 1; i >= curSegIdx; i--) {
                 lanes_[i] = lanes_[oriSize--];
             }
-            for (size_t i = start; i < curSegIdx; ++i) {
+            for (size_t i = static_cast<size_t>(start); i < curSegIdx; ++i) {
                 lanes_[i] = std::vector<Lane>(sections[i].crossCount.value_or(1));
             }
         }
